+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgload (development version)
 
+* `load_all()` now issues a warning if exported objects conflict with objects defined in the global environment (#112)
+
 * `dev_example()` now works after removing an inconsistent call to `load_all()` (@riccardoporreca, #122).
 
 # pkgload 1.0.2